Add CL_DEVICE_DOUBLE_FP_CONFIG
Note that "cl.h" defines the macro only for OpenCL 1.2 or greater since  
"cl_khr_fp64" was made a core-optional feature from OpenCL 1.2 onwards.  
This creates an issue when users targeting OpenCL 1.1 or lower try to query  
the device for double precision support using the macro.  

This change fixes this issue by adding the macro in "cl_ext.h" for OpenCL 1.1 or lower.