sort new extensions programmatically
The generated extension headers maintained an ordered list of older
extensions to minimize diffs with the non-generated extension headers,
but there is no reason to require this for new extensions.  If a new
extension is not found in the list, sort it at the end automatically.
Note, Khronos (KHR) extensions will also be sorted before multi-vendor
(EXT) extensions, and before vendor extensions.